Export module into umd format
This removes the browser.js file and uses the built in feature of
browserify to create a module in umd format which makes it usable in
node, with require js or in the browser without any module system.
Source maps has also been added to the build folder.

Section files have been added package.json to make sure everything is
included in the deployed package.

Also: uglify has been replaced with the more mature uglify-js